migrate tests from existing rules
